refactor out invite types module
