+++ services/user/Invite/plugin/src/lib.rs
@@ -1,6 +1,8 @@
 #[allow(warnings)]
 mod bindings;
-use base64::{engine::general_purpose::URL_SAFE, Engine};
+mod errors;
+mod types;
+
 use bindings::accounts::plugin as Accounts;
 use bindings::auth_sig::plugin::{keyvault, types::Pem};
 use bindings::exports::invite;
@@ -8,15 +10,12 @@ use bindings::host::common::{client as Client, server as Server, types as Common
 use bindings::invite::plugin::types::{Invite, InviteId, InviteState, Url};
 use bindings::transact::plugin::intf as Transact;
 use chrono::DateTime;
+use errors::ErrorType::*;
 use fracpack::Pack;
 use invite::plugin::{invitee::Guest as Invitee, inviter::Guest as Inviter};
 use psibase::services::invite as InviteService;
+use types::*;
 use CommonTypes::OriginationData;
-
-use serde::{Deserialize, Serialize};
-
-mod errors;
-use errors::ErrorType::*;
 
 /* TODO:
     /// This doesn't need to be exposed, it can just be jammed into various plugin functions
@@ -26,78 +25,9 @@ use errors::ErrorType::*;
     void delExpired(uint32_t maxDeleted);
 */
 
-#[derive(Serialize, Deserialize)]
-struct InviteParams {
-    app: String,
-    pk: String,
-    cb: String,
-}
+struct InvitePlugin;
 
-#[derive(Deserialize)]
-struct ResponseRoot<T> {
-    data: T,
-}
-trait TryParseGqlResponse: Sized {
-    fn from_gql(s: String) -> Result<Self, CommonTypes::Error>;
-}
-
-//#[qgl_query(name="getInvite")] // <-- Todo: Create this procedural macro ...
-#[allow(non_snake_case)]
-#[derive(Deserialize)]
-struct InviteRecordSubset {
-    inviter: psibase::AccountNumber,
-    actor: psibase::AccountNumber,
-    expiry: u32,
-    state: u8,
-}
-
-//                                           ...Such that something like the below code is generated:
-#[allow(non_snake_case)]
-#[derive(Deserialize)]
-struct GetInviteResponse {
-    getInvite: Option<InviteRecordSubset>,
-}
-impl TryParseGqlResponse for InviteRecordSubset {
-    fn from_gql(response: String) -> Result<Self, CommonTypes::Error> {
-        let response_root: ResponseRoot<GetInviteResponse> =
-            serde_json::from_str(&response).map_err(|e| QueryError.err(&e.to_string()))?;
-        Ok(response_root.data.getInvite.ok_or_else(|| {
-            QueryError.err("Unable to extract InviteRecordSubset from query response")
-        })?)
-    }
-}
-//                                            />
-
-struct Component;
-
-impl From<InviteParams> for InviteId {
-    fn from(params: InviteParams) -> Self {
-        let params_str = serde_json::to_string(&params).unwrap();
-        URL_SAFE.encode(params_str)
-    }
-}
-
-trait TryFromInviteId: Sized {
-    fn try_from_invite_id(id: InviteId) -> Result<Self, CommonTypes::Error>;
-}
-
-impl TryFromInviteId for InviteParams {
-    fn try_from_invite_id(id: InviteId) -> Result<Self, CommonTypes::Error> {
-        let bytes = URL_SAFE
-            .decode(id.to_owned())
-            .map_err(|_| DecodeInviteError.err("Error decoding base64"))?;
-
-        let str = String::from_utf8(bytes)
-            .map_err(|_| DecodeInviteError.err("Error converting from UTF8"))?;
-
-        let result: InviteParams = serde_json::from_str(&str)
-            .map_err(|_| DecodeInviteError.err("Error deserializing JSON string into object"))?;
-
-        Ok(result)
-    }
-}
-
-impl Invitee for Component {
+impl Invitee for InvitePlugin {
     fn accept_with_new_account(account: String, id: InviteId) -> Result<(), CommonTypes::Error> {
         let accepted_by = psibase::AccountNumber::from_exact(&account).or_else(|_| {
             return Err(InvalidAccount.err(&account));
@@ -166,7 +96,7 @@ impl Invitee for Component {
     }
 }
 
-impl Inviter for Component {
+impl Inviter for InvitePlugin {
     fn generate_invite(callback_subpath: String) -> Result<Url, CommonTypes::Error> {
         let keypair = keyvault::generate_unmanaged_keypair()?;
 
@@ -206,4 +136,4 @@ impl Inviter for Component {
     }
 }
 
-bindings::export!(Component with_types_in bindings);
+bindings::export!(InvitePlugin with_types_in bindings);
